fix profile reducer for name

+++ client/src/App.js
@@ -22,7 +22,7 @@ import AddListingContainer from "./containers/AddListingContainer";
 import Notifier from "./containers/Notifier";
 import ContentLibrary from "./containers/ContentLibrary";
 import Spinner from "./components/Spinner";
-import UpdateUser from "./components/UpdateUser";
+import UpdateUserContainer from "./containers/UpdateUserContainer";
 import ListingsMap from "./containers/ListingsMap";
 
 import Hero from "./img/hero_keys_1920x1080.jpg";
@@ -83,6 +83,11 @@ const styles = theme => ({
     height: 100,
     margin: "20px auto",
     width: 100
+  },
+  buttonWrap: {
+    width: "100vw",
+    display: "flex",
+    justifyContent: "center"
   },
   footer: {
     width: "100vw",
@@ -176,22 +181,68 @@ export class AppUnconnected extends Component {
     // If not logged in, check local storage for authToken
     // if it doesn't exist, it returns the string "undefined"
     if (!this.props.appState.loggedIn) {
-      const authToken = window.localStorage.getItem("authToken");
-      const userId = window.localStorage.getItem("userId");
-      if (
-        authToken &&
-        authToken !== "undefined" &&
-        userId &&
-        userId !== "undefined"
-      ) {
-        this.props.apiProfile
-          .validateToken(authToken, userId)
-          .then(result => {
-            if (result.type === "VALIDATE_TOKEN_FAILURE") {
-              window.localStorage.clear();
-            }
-          })
-          .catch(err => console.log(err));
+      // don't run this sequence if landing on admin dash for first time
+      // after google auth -- there will be nothing in localstorage yet
+      if (!(this.props.match && this.props.match.params.id)) {
+        // console.log("not logged in, looking for id & token in localStorage");
+        const authToken = window.localStorage.getItem("authToken");
+        const userId = window.localStorage.getItem("userId");
+        // console.log(`authToken: ${!!authToken}, userId: ${userId}`);
+        if (
+          authToken &&
+          authToken !== "undefined" &&
+          userId &&
+          userId !== "undefined"
+        ) {
+          // console.log("found id & token in localstorage, validating token");
+          // console.log(!!authToken, userId);
+          this.props.apiProfile
+            .validateToken(authToken, userId)
+            .then(result => {
+              // console.log(result.type);
+              if (result.type === "VALIDATE_TOKEN_FAILURE") {
+                // console.log("VALIDATE_TOKEN_FAILURE: clearing localStorage");
+                return window.localStorage.clear();
+              }
+              if (
+                result.type === "VALIDATE_TOKEN_SUCCESS" &&
+                authToken &&
+                authToken !== "undefined" &&
+                userId &&
+                userId !== "undefined"
+              ) {
+                // console.log(
+                //   `validate token success: ${!!authToken}, ${userId}`
+                // );
+                this.props.apiProfile
+                  .getProfile(authToken, userId)
+                  .then(result => {
+                    // console.log(result.type);
+                    if (result.type === "GET_PROFILE_SUCCESS") {
+                      // console.log(
+                      //   `setting user type here: ${result.payload.type}`
+                      // );
+                      this.props.actions.setLoggedIn(result.payload.type);
+                      // check for redirect url in local storage
+                      const redirect = window.localStorage.getItem("redirect");
+                      if (redirect) {
+                        // redirect to originally requested page and then
+                        // clear value from local storage
+                        this.props.history.push(redirect);
+                        window.localStorage.removeItem("redirect");
+                      }
+                    } else {
+                      // console.log("not logged in", authToken, userId);
+                      // console.log(result.type);
+                    }
+                  });
+              }
+            })
+            .catch(err => {
+              console.log(err);
+              return window.localStorage.clear();
+            });
+        }
       }
     }
   }
@@ -256,7 +307,10 @@ export class AppUnconnected extends Component {
             <Route
               path="/user"
               render={routeProps => (
-                <UpdateUser setRedirect={this.setRedirect} {...routeProps} />
+                <UpdateUserContainer
+                  setRedirect={this.setRedirect}
+                  {...routeProps}
+                />
               )}
             />
             <Route
